FLYW-56 move: reservation 탑승객 도메인 파일 passenger 패키지로 이동 (#22)
