+++ src/main/java/com/flyway/passenger/dto/BaggageSaveRequest.java
@@ -1,4 +1,4 @@
-package com.flyway.reservation.domain;
+package com.flyway.passenger.dto;
 
 import lombok.*;
