+++ src/main/java/com/flyway/passenger/dto/PassengerView.java
@@ -1,26 +1,28 @@
-package com.flyway.reservation.dto;
+package com.flyway.passenger.dto;
 
 import lombok.*;
+
+import java.time.LocalDate;
 
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class PassengerSaveRow {
+public class PassengerView {
 
-    private String passengerId; // hidden (있으면 update, 없으면 insert)
+    private String passengerId;
     private String krFirstName;
     private String krLastName;
     private String firstName;
     private String lastName;
-    private String birth;   // yyyy-MM-dd (문자열 -> 서비스에서 LocalDate )
-    private String gender;  // M/F
+    private LocalDate birth;     // DATE
+    private String gender;       // M/F
     private String email;
     private String phoneNumber;
 
     private String passportNo;
     private String country;
-    private String passportExpiryDate;     // yyyy-MM-dd
+    private LocalDate passportExpiryDate;
     private String passportIssueCountry;
 }
